experimental.9

+++ src/main/java/com/github/winterreisender/webviewko/WebviewJNA.kt
@@ -1,7 +1,21 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
+/*
+ * Copyright (c) 2022  Winterreisender
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX short identifier: **Apache-2.0**
+ */
+
 package com.github.winterreisender.webviewko
 
 import com.sun.jna.*
@@ -88,15 +102,15 @@ object WebviewJNA {
         return WebviewLibrary.INSTANCE
     }
 
-    @Deprecated("Only For Test")
-    fun getRawInstance() = WebviewLibrary.INSTANCE
-
+    // Copy WebView2Loader.dll in Windows
     fun extractDependencies() {
-        if (System.getProperty("os.name").lowercase().contains("win"))
+        val targetPath = Path.of("${System.getProperty("user.dir")}/WebView2Loader.dll")
+
+        if (System.getProperty("os.name").lowercase().contains("win") && Files.notExists(targetPath))
             try {
                 Files.copy(
                     this::class.java.classLoader.getResourceAsStream("WebView2Loader.dll")!!,
-                    Path.of("${System.getProperty("user.dir")}/WebView2Loader.dll")
+                    targetPath
                 )
             }catch (e :FileAlreadyExistsException) {
                 println("File Already Exists")
